pyproject.toml
[project]
name = "Pokedex"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "ositoMalvado", email = "mordecaaii@gmail.com" }
]
dependencies = [
  "aiohttp>=3.9.5",
  "flet==0.26.0",
  "pillow==10.4.0",
]

[tool.flet]
# org name in reverse domain name notation, e.g. "com.mycompany".
# Combined with project.name to build bundle ID for iOS and Android apps
org = "com.ositoMalvado"

# project display name that is used as an app title on Android and iOS home screens,
# shown in window titles and about app dialogs on desktop.
product = "Pokedex"

# company name to display in about app dialogs
company = "ositoMalvado"

# copyright text to display in about app dialogs
copyright = "Copyright (C) 2025 by ositoMalvado"

[tool.flet.app]
path = ""

[tool.uv]
dev-dependencies = [
    "flet[all]==0.26.0",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
flet = {extras = ["all"], version = "0.26.0"}
